+++ run_mac.sh
@@ -1,7 +1,7 @@
 #!/bin/zsh
 
 n_cores=6
-n_events=50000
+n_events=10000
 n_runs=7
 
 #beam properties
@@ -19,7 +19,7 @@ angle=PHI                       #in all caps
 
 # smearing=0.193
 # smearing=0.0
-smearing=0.21
+smearing=0.23
 
 rm -r output/*.bak
 rm -r output/${particle}*-*
@@ -29,6 +29,9 @@ sed -i .bak "s+/run/beamOn.*+/run/beamOn $n_events+" n_event.mac
 #echo -e "\e[31mNumber of events: $n_events \e[39m..."
 sed -i .bak "s/G4int NoE = .*/G4int NoE = $n_events;/" ../JEPO-1.1/src/BT2017PriGenAct.cc
 sed -i .bak "s/G4double Smear = .*/G4double Smear = $smearing;/" ../JEPO-1.1/src/BT2017EveAct.cc
+
+sed -i .bak "s+//comment this section...+//comment this section... \n/*+" output/Simulation_runner.cpp
+sed -i .bak "s+//... if ROOT is installed from source+*/\n//... if ROOT is installed from source+" output/Simulation_runner.cpp
 
 sed -i .bak "s/^const int n_runs.*/const int n_runs = $n_runs;/" output/Simulation_runner.cpp
 sed -i .bak "s/^const int nth.*/const int nth = $n_cores;/" output/Simulation_runner.cpp
@@ -62,9 +65,11 @@ fi
 if [[ "$configuration" == "PARALLEL" ]]
 then
   sed -i .bak "s/^setting=.*/setting='G4'/" output/Peak_fitter_2.py
+  root -q -l "beam_profile.cpp($n_events)" # Own code for generating beam profile, comment for new users
+  sed -i .bak "s/m_FlagBeamFile =.*/m_FlagBeamFile = 1;/" ../JEPO-1.1/src/BT2017PriGenAct.cc # Own code for generating beam profile, comment for new users (or set = 0)
 elif [[ "$configuration" == "PERPENDICULAR" ]]
 then
-  sed -i .bak "s/G4String angle=.*/\tG4String angle=\"$angle\";/" ../JEPO-1.1/src/BT2017PriGenAct.cc
+  sed -i .bak "s/G4String angle=.*/G4String angle=\"$angle\";/" ../JEPO-1.1/src/BT2017PriGenAct.cc
 fi
 
 make clean
@@ -75,6 +80,7 @@ subtracted=(0.24739106414386924 5.065845196500294 10.152798578987705 15.21887072
 
 # phi_array=(0 45 90 135 225 270 315)
 phi_array=(0.0 22.5 45.0 67.5 90.0 112.5 135.0 157.5 202.5 225.0 247.5 270.0 292.5 315.0 337.5) #15
+
 
 for (( i=0; i<$n_runs; i++ ))
 do
